Mise à jour Season

+++ Season/main.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main() {
    const Season& season = Season::Spring;
-   for (int i = 0; i < Season::size(); i++)
+   for (size_t i = 0; i < Season::size(); i++)
       cout << Season::get(i).name() << endl;
    for (const Season* s = *Season::values();
         s <= Season::values()[Season::size() - 1]; s++)
